Add ability to use docker based build for insource-builds (#891)
- adds a step to build builder which can be run with the DALI
  source code mounted, so consecutive builds are faster
- adds rebuilding of docker images optional when they don't
  exists or user forces their rebuild
- adds the ability to customize build directory by the option, so it
  is easier to build different CUDA and Python versions next
  to each other
- update devtools to version 4 (gcc 5.3.1_ in manylinux3 image
  (newer are incompatible with CUDA9)

+++ docker/build.sh
@@ -1,4 +1,37 @@
 #!/bin/bash -xe
+
+usage="ENV1=VAL1 ENV2=VAL2 [...] $(basename "$0") [-h] -- this is simple, one click build script for DALI utilizing Docker as
+a build environment
+
+To change build configuration please export appropriate env variables (for exact meaning please check the README):
+PYVER=[default 3.7]
+CUDA_VERSION=[default 10, accepts also 9]
+NVIDIA_BUILD_ID=[default 12345]
+CREATE_WHL=[default YES]
+CREATE_RUNNER=[default NO]
+DALI_BUILD_FLAVOR=[default is empty]
+CMAKE_BUILD_TYPE=[default is Release]
+BUILD_INHOST=[create build dir with object outside docker, just mount it as a volume]
+REBUILD_BUILDERS=[default is NO]
+REBUILD_MANYLINUX=[default is NO]
+DALI_BUILD_DIR=[default is build-docker-\${CMAKE_BUILD_TYPE}-\${PYV}-\${CUDA_VERSION}]
+
+where:
+    -h  show this help text"
+
+while getopts 'h' option; do
+  case "$option" in
+    h) echo "$usage"
+       exit
+       ;;
+   \?) printf "illegal option: -%s\n" "$OPTARG" >&2
+       echo "$usage" >&2
+       exit 1
+       ;;
+  esac
+done
+shift $((OPTIND - 1))
+
 #########Set Me###############
 export PYVER=${PYVER:-3.5}
 export PYV=${PYVER/./}
@@ -7,10 +40,15 @@ export NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID:-12345}
 export CREATE_WHL=${CREATE_WHL:-YES}
 export CREATE_RUNNER=${CREATE_RUNNER:-NO}
 export DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR}
-export BUILD_TYPE=${BUILD_TYPE:-Release}
+export CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}
+export BUILD_INHOST=${BUILD_INHOST:-YES}
+export REBUILD_BUILDERS=${REBUILD_BUILDERS:-NO}
+export REBUILD_MANYLINUX=${REBUILD_MANYLINUX:-NO}
+export DALI_BUILD_DIR=${DALI_BUILD_DIR:-build-docker-${CMAKE_BUILD_TYPE}-${PYV}-${CUDA_VERSION}}
 #################################
 export DEPS_IMAGE=dali_cu${CUDA_VERSION}.deps
 export BUILDER=dali_${PYV}_cu${CUDA_VERSION}.build
+export BUILDER_WHL=dali_${PYV}_cu${CUDA_VERSION}.build_w_whl
 export RUN_IMG=dali_${PYV}_cu${CUDA_VERSION}.run
 export GIT_SHA=$(git rev-parse HEAD)
 export DALI_TIMESTAMP=$(date +%Y%m%d)
@@ -23,21 +61,67 @@ then
     exit 1
 fi
 
-# build manylinux3
-pushd ../third_party/manylinux/
-git checkout 96b47a25673b33c728e49099a3a6b1bf503a18c2 || echo -e "Did you forget to \`git clone --recursive\`? Try this:\n" \
-                                                                 "  git submodule sync --recursive && \n" \
-                                                                 "  git submodule update --init --recursive && \n"
-git am ../../docker/0001-An-approximate-manylinux3.patch
-PLATFORM=$(uname -m) TRAVIS_COMMIT=latest ./build.sh
-popd
+# build manylinux3 if needed
+if [[ "$(docker images -q manylinux3_x86_64 2> /dev/null)" == "" || "$REBUILD_MANYLINUX" != "NO" ]]; then
+    pushd ../third_party/manylinux/
+    git checkout 96b47a25673b33c728e49099a3a6b1bf503a18c2 || echo -e "Did you forget to \`git clone --recursive\`? Try this:\n" \
+                                                                    "  git submodule sync --recursive && \n" \
+                                                                    "  git submodule update --init --recursive && \n"
+    git am ../../docker/0001-An-approximate-manylinux3.patch
+    PLATFORM=$(uname -m) TRAVIS_COMMIT=latest ./build.sh
+    popd
+fi
 
 pushd ../
-docker build -t ${DEPS_IMAGE} --build-arg "FROM_IMAGE_NAME"=manylinux3_x86_64 --build-arg "USE_CUDA_VERSION=${CUDA_VERSION}" -f Dockerfile.deps .
-echo "Build image:" ${BUILDER}
-docker build -t ${BUILDER} --build-arg "DEPS_IMAGE_NAME=${DEPS_IMAGE}" --build-arg "PYVER=${PYVER}" --build-arg "PYV=${PYV}" --build-arg "NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID}" \
+# build deps image if needed
+if [[ "$(docker images -q ${DEPS_IMAGE} 2> /dev/null)" == "" || "$REBUILD_BUILDERS" != "NO" ]]; then
+    echo "Build deps: " ${DEPS_IMAGE}
+    docker build -t ${DEPS_IMAGE} --build-arg "FROM_IMAGE_NAME"=manylinux3_x86_64 --build-arg "USE_CUDA_VERSION=${CUDA_VERSION}" -f Dockerfile.deps .
+fi
+
+# build builder image if needed
+if [[ "$(docker images -q ${BUILDER} 2> /dev/null)" == "" || "$REBUILD_BUILDERS" != "NO" ]]; then
+    echo "Build light image:" ${BUILDER}
+    docker build -t ${BUILDER} --build-arg "DEPS_IMAGE_NAME=${DEPS_IMAGE}" --build-arg "PYVER=${PYVER}" --build-arg "PYV=${PYV}" --build-arg "NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID}" \
                            --build-arg "NVIDIA_DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR}" --build-arg "GIT_SHA=${GIT_SHA}" --build-arg "DALI_TIMESTAMP=${DALI_TIMESTAMP}" \
-                           --build-arg "CMAKE_BUILD_TYPE=${BUILD_TYPE}" .
+                           --build-arg "CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}" --target builder .
+fi
+
+if [ "$BUILD_INHOST" = "YES" ]; then
+    # build inside the source tree
+    docker run --rm -u 1000:1000 -v $(pwd):/opt/dali ${BUILDER} /bin/bash -c "mkdir -p /opt/dali/${DALI_BUILD_DIR} && \
+                                        cd /opt/dali/${DALI_BUILD_DIR} &&         \
+                                        rm -rf /opt/dali/${DALI_BUILD_DIR}/nvidia* && \
+                                        CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}      \
+                                        BUILD_TEST=${BUILD_TEST}                  \
+                                        BUILD_BENCHMARK=${BUILD_BENCHMARK}        \
+                                        BUILD_NVTX=${BUILD_NVTXBUILD_NVTX}        \
+                                        BUILD_PYTHON=${BUILD_PYTHON}              \
+                                        BUILD_LMDB=${BUILD_LMDB}                  \
+                                        BUILD_TENSORFLOW=${BUILD_TENSORFLOW}      \
+                                        BUILD_JPEG_TURBO=${BUILD_JPEG_TURBO}      \
+                                        BUILD_NVJPEG=${BUILD_NVJPEG}              \
+                                        BUILD_NVOF=${BUILD_NVOF}                  \
+                                        BUILD_NVDEC=${BUILD_NVDEC}                \
+                                        BUILD_NVML=${BUILD_NVML}                  \
+                                        NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID}        \
+                                        GIT_SHA=${GIT_SHA}                        \
+                                        DALI_TIMESTAMP=${DALI_TIMESTAMP}          \
+                                        NVIDIA_DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR} \
+                                        /opt/dali/docker/build_helper.sh &&       \
+                                        rm -rf /opt/dali/${DALI_BUILD_DIR}/nvidia* && \
+                                        cp /wheelhouse/* ./"
+    if [ "$CREATE_WHL" = "YES" ]; then
+        mkdir -p ./wheelhouse
+        cp $(pwd)/${DALI_BUILD_DIR}/nvidia* ./wheelhouse/
+    fi
+else
+    echo "Build image:" ${BUILDER_WHL}
+    docker build -t ${BUILDER_WHL} --build-arg "DEPS_IMAGE_NAME=${DEPS_IMAGE}" --build-arg "PYVER=${PYVER}" --build-arg "PYV=${PYV}" --build-arg "NVIDIA_BUILD_ID=${NVIDIA_BUILD_ID}" \
+                            --build-arg "NVIDIA_DALI_BUILD_FLAVOR=${DALI_BUILD_FLAVOR}" --build-arg "GIT_SHA=${GIT_SHA}" --build-arg "DALI_TIMESTAMP=${DALI_TIMESTAMP}" \
+                            --build-arg "CMAKE_BUILD_TYPE=${BUILD_TYPE}" --cache-from "${BUILDER}" .
+fi
+
 
 if [ "$CREATE_RUNNER" = "YES" ]; then
     echo "Runner image:" ${RUN_IMG}
@@ -53,15 +137,33 @@ if [ "$CREATE_RUNNER" = "YES" ]; then
     fi
     echo ${CUDA_VERSION}
     echo ${CUDA_IMAGE_NAME}
-
-    docker build -t ${RUN_IMG} --build-arg "BUILD_IMAGE_NAME=${BUILDER}" --build-arg "CUDA_IMAGE_NAME=${CUDA_IMAGE_NAME}" --build-arg "PYVER=${PYVER}" --build-arg "PYV=${PYV}" -f Docker_run_cuda .
+    export BUILDER_TMP=${BUILDER_WHL}
+    # for intree build we don't have docker image with whl inside so create one
+    if [ "$BUILD_INHOST" = "YES" ]; then
+        export BUILDER_TMP=${BUILDER}_tmp
+        DOCKER_FILE_TMP=$(mktemp)
+        echo -e "FROM scratch\n"          \
+                "COPY ./${DALI_BUILD_DIR}/nvidia* /wheelhouse/\n" > ${DOCKER_FILE_TMP}
+        docker build -t ${BUILDER_TMP} -f ${DOCKER_FILE_TMP} .
+        rm ${DOCKER_FILE_TMP}
+    fi
+    docker build -t ${RUN_IMG} --build-arg "BUILD_IMAGE_NAME=${BUILDER_TMP}" --build-arg "CUDA_IMAGE_NAME=${CUDA_IMAGE_NAME}" --build-arg "PYVER=${PYVER}" --build-arg "PYV=${PYV}" -f Docker_run_cuda .
+    # remove scratch image
+    if [ "$BUILD_INHOST" = "YES" ]; then
+        docker rmi ${BUILDER_TMP}
+    fi
 fi
 
 if [ "$CREATE_WHL" = "YES" ]; then
-    export CONTAINER="extract-tmp"
-    docker create --name "${CONTAINER}" ${BUILDER}
-    rm -rf ./wheelhouse
-    docker cp "${CONTAINER}:/wheelhouse/" "./"
-    docker rm -f "${CONTAINER}"
+    if [ "$BUILD_INHOST" = "YES" ]; then
+        mkdir -p ./wheelhouse
+        cp $(pwd)/${DALI_BUILD_DIR}/nvidia* ./wheelhouse/
+    else
+        export CONTAINER="extract-tmp"
+        docker create --name "${CONTAINER}" ${BUILDER_WHL}
+        rm -rf ./wheelhouse
+        docker cp "${CONTAINER}:/wheelhouse/" "./"
+        docker rm -f "${CONTAINER}"
+    fi
 fi
 popd
